+++ tests/testthat/test-lp_nl_panel.R
@@ -701,9 +701,46 @@ set.seed(123)
   })
 
 
+  test_that("Test that shock name is correct", {
+
+
+    testthat::expect_error(lp_nl_panel(data_set          = data_set,
+                                       data_sample       = 'Full',
+                                       endog_data        = "y",
+                                       cumul_mult        = T,
+
+                                       shock             = "shock",
+                                       diff_shock        = F,
+                                       panel_model       = "within",
+                                       panel_effect      = "individual",
+                                       robust_cov        = "Vcx",
+
+                                       switching         = "x_1",
+                                       lag_switching     = T,
+                                       use_logistic      = T,
+                                       use_hp            = T,
+                                       lambda            = 3,
+                                       gamma             = 3,
+
+                                       use_gmm           = F,
+
+                                       c_exog_data       = colnames(data_set)[4:6],
+                                       l_exog_data       = colnames(data_set)[4:6],
+                                       lags_exog_data    = 2,
+                                       c_fd_exog_data    = colnames(data_set)[4:6],
+                                       l_fd_exog_data    = colnames(data_set)[4:6],
+                                       lags_fd_exog_data = 2,
+
+                                       confint           = 1,
+                                       hor               = 2),
+                           'Please use another name for your shock variable".
+         Your current name would lead to a naming problem during estimation.', fixed = TRUE)
+  })
+
+
 test_that("Test whether output coefficients from 'lp_nl_panel_results' coincide
             with coefficients estimated by using the matrix prepared in 'lp_nl_panel'", {
-
+              suppressWarnings(
               # Estimate panel model
               results_panel <-  lp_nl_panel(data_set          = data_set,
                                             data_sample       = 'Full',
@@ -730,7 +767,7 @@ test_that("Test whether output coefficients from 'lp_nl_panel_results' coincide
                                             lags_fd_exog_data = 2,
 
                                             confint           = 1.67,
-                                            hor               = 10)
+                                            hor               = 10))
 
 
               lp_nl_panel_results <- results_panel$reg_summaries[[1]]
@@ -898,8 +935,322 @@ test_that("Test gmm model", {
                                                   lags_fd_exog_data = NULL,
 
                                                   confint           = 1.67,
-                                                  hor               = 10),
+                                                  hor               = 3),
                                       NA)
             })
 
+
+
+test_that("Check output of switching variable I", {
+
+
+  # Estimate panel model
+  results_nl <- lp_nl_panel(data_set                   = data_set,
+                                     data_sample       = 'Full',
+                                     endog_data        = "y",
+                                     cumul_mult        = TRUE,
+
+                                     shock             = "x_1",
+                                     diff_shock        = FALSE,
+                                     panel_model       = "within",
+                                     panel_effect      = "individual",
+                                     robust_cov        = "vcovSCC",
+
+                                     switching         = "x_1",
+                                     lag_switching     = FALSE,
+                                     use_logistic      = FALSE,
+                                     use_hp            = FALSE,
+                                     lambda            = 3,
+                                     gamma             = 3,
+
+                                     use_gmm           = F,
+
+                                     c_exog_data       = NULL,
+                                     l_exog_data       = "y",
+                                     lags_exog_data    = 1,
+                                     c_fd_exog_data    = NULL,
+                                     l_fd_exog_data    = NULL,
+                                     lags_fd_exog_data = NULL,
+
+                                     confint           = 1.67,
+                                     hor               = 3)
+
+  # Test whether shock variable is correct
+  testthat::expect_equal(results_nl$fz$fz[, 1], data_set$x_1)
+
+
+
+  # States from output
+  shock_s1_output <- results_nl$xy_data_sets[[1]]$shock_s1
+  shock_s2_output <- results_nl$xy_data_sets[[1]]$shock_s2
+
+  # Make lagged tibble based on data set
+  data_set_lag <- data_set %>%
+                  arrange(cross_section, time_section) %>%
+                  group_by(cross_section)              %>%
+                  mutate_at(vars(y), funs(dplyr::lag(., 1))) %>%
+                  ungroup()
+
+  lag_indicator <-  !is.na(data_set_lag$y)
+
+  # Manual states
+  shock_s1_manual <- data_set$x_1*(1  - data_set$x_1)
+  shock_s1_manual <- shock_s1_manual[lag_indicator]
+
+  shock_s2_manual <- data_set$x_1*(data_set$x_1)
+  shock_s2_manual <- shock_s2_manual[lag_indicator]
+
+  # Test whether states of shock variable are correctly estimated
+  testthat::expect_equal(as.numeric(shock_s1_output), shock_s1_manual)
+  testthat::expect_equal(as.numeric(shock_s2_output), shock_s2_manual)
+
+
+})
+
+
+
+test_that("Check output of switching variable II", {
+
+
+
+  # Estimate panel model
+  results_nl <- lp_nl_panel(data_set          = data_set,
+                            data_sample       = 'Full',
+                            endog_data        = "y",
+                            cumul_mult        = TRUE,
+
+                            shock             = "x_1",
+                            diff_shock        = FALSE,
+                            panel_model       = "within",
+                            panel_effect      = "individual",
+                            robust_cov        = "vcovSCC",
+
+                            switching         = "x_1",
+                            lag_switching     = TRUE,
+                            use_logistic      = FALSE,
+                            use_hp            = FALSE,
+                            lambda            = 3,
+                            gamma             = 3,
+
+                            use_gmm           = F,
+
+                            c_exog_data       = NULL,
+                            l_exog_data       = "y",
+                            lags_exog_data    = 1,
+                            c_fd_exog_data    = NULL,
+                            l_fd_exog_data    = NULL,
+                            lags_fd_exog_data = NULL,
+
+                            confint           = 1.67,
+                            hor               = 3)
+
+  lag_dplyr  <- data_set %>%
+                group_by(cross_section) %>%
+                           mutate_at(vars(x_1), funs(dplyr::lag(.,1))) %>%
+                           ungroup() %>%
+                           select(x_1)  %>%
+                           as.matrix()  %>%
+                           as.numeric()
+
+  fz_output  <- results_nl$fz$fz
+
+  # Test whether switching variable is correctly specified
+  testthat::expect_equal(lag_dplyr, fz_output)
+
+
+  # Test whether states of shock variable are correctly estimated
+  # States from output
+  shock_s1_output <- results_nl$xy_data_sets[[1]]$shock_s1
+  shock_s2_output <- results_nl$xy_data_sets[[1]]$shock_s2
+
+
+  # Manual states
+  shock_s1_manual <- as.numeric(na.omit(data_set$x_1*(1  - lag_dplyr)))
+  shock_s2_manual <- as.numeric(na.omit(data_set$x_1*lag_dplyr))
+
+  # Test whether states of shock variable are correctly specified
+  testthat::expect_equal(as.numeric(shock_s1_output), shock_s1_manual)
+  testthat::expect_equal(as.numeric(shock_s2_output), shock_s2_manual)
+
+
+})
+
+
+
+
+
+test_that("Check output of switching variable III", {
+
+  # Estimate panel model
+  results_nl <- lp_nl_panel(data_set                   = data_set,
+                            data_sample       = 'Full',
+                            endog_data        = "y",
+                            cumul_mult        = TRUE,
+
+                            shock             = "x_1",
+                            diff_shock        = FALSE,
+                            panel_model       = "within",
+                            panel_effect      = "individual",
+                            robust_cov        = "vcovSCC",
+
+                            switching         = "x_1",
+                            lag_switching     = FALSE,
+                            use_logistic      = TRUE,
+                            use_hp            = FALSE,
+                            lambda            = 3,
+                            gamma             = 3,
+
+                            use_gmm           = F,
+
+                            c_exog_data       = NULL,
+                            l_exog_data       = "y",
+                            lags_exog_data    = 1,
+                            c_fd_exog_data    = NULL,
+                            l_fd_exog_data    = NULL,
+                            lags_fd_exog_data = NULL,
+
+                            confint           = 1.67,
+                            hor               = 3)
+
+  # Use logistic function
+  logistic_function <- function(z_0){
+
+    switching_val <- exp(-3*z_0)/
+                    (1 + exp(-3*z_0))
+
+  }
+
+
+  # Compute manual values of switching variable
+  logistic_dplyr  <- data_set %>%
+                    group_by(cross_section) %>%
+                    mutate_at(vars(x_1), funs(logistic_function(.))) %>%
+                    ungroup() %>%
+                    select(x_1)  %>%
+                    as.matrix()  %>%
+                    as.numeric()
+
+  fz_output  <- results_nl$fz$fz
+
+  # Test whether switching variable from function and manual values are identical
+  testthat::expect_equal(logistic_dplyr, fz_output)
+
+
+  # States from output
+  shock_s1_output <- results_nl$xy_data_sets[[1]]$shock_s1
+  shock_s2_output <- results_nl$xy_data_sets[[1]]$shock_s2
+
+  # Make lagged tibble based on data set
+  data_set_lag <- data_set %>%
+                  arrange(cross_section, time_section) %>%
+                  group_by(cross_section)              %>%
+                  mutate_at(vars(y), funs(dplyr::lag(., 1))) %>%
+                  ungroup()
+
+  lag_indicator <-  !is.na(data_set_lag$y)
+
+  # Manual states
+  shock_s1_manual <- data_set$x_1*(1  - logistic_dplyr)
+  shock_s1_manual <- shock_s1_manual[lag_indicator]
+
+  shock_s2_manual <- data_set$x_1*(logistic_dplyr)
+  shock_s2_manual <- shock_s2_manual[lag_indicator]
+
+  # Test whether states of shock variable are correctly estimated
+  testthat::expect_equal(as.numeric(shock_s1_output), shock_s1_manual)
+  testthat::expect_equal(as.numeric(shock_s2_output), shock_s2_manual)
+
+
+})
+
+
+
+
+test_that("Check output of switching variable IV", {
+
+
+  # Estimate panel model
+  results_nl <- lp_nl_panel(data_set                   = data_set,
+                            data_sample       = 'Full',
+                            endog_data        = "y",
+                            cumul_mult        = TRUE,
+
+                            shock             = "x_1",
+                            diff_shock        = FALSE,
+                            panel_model       = "within",
+                            panel_effect      = "individual",
+                            robust_cov        = "vcovSCC",
+
+                            switching         = "x_1",
+                            lag_switching     = TRUE,
+                            use_logistic      = TRUE,
+                            use_hp            = FALSE,
+                            lambda            = 3,
+                            gamma             = 3,
+
+                            use_gmm           = F,
+
+                            c_exog_data       = NULL,
+                            l_exog_data       = "y",
+                            lags_exog_data    = 1,
+                            c_fd_exog_data    = NULL,
+                            l_fd_exog_data    = NULL,
+                            lags_fd_exog_data = NULL,
+
+                            confint           = 1.67,
+                            hor               = 3)
+
+  # Logistic function
+  logistic_function <- function(z_0){
+
+    switching_val <- exp(-3*z_0)/
+      (1 + exp(-3*z_0))
+
+  }
+
+
+  # Compute manual switching values based on logistic function
+  fz_dplyr  <- data_set %>%
+                group_by(cross_section) %>%
+                mutate_at(vars(x_1), funs(logistic_function(.))) %>%
+                mutate_at(vars(x_1), funs(dplyr::lag(., 1)))     %>%
+                ungroup() %>%
+                select(x_1)  %>%
+                as.matrix()  %>%
+                as.numeric()
+
+  # Switching values from output
+  fz_output  <- results_nl$fz$fz
+
+  # Test whether switching variable from output and manual are identical
+  testthat::expect_equal(fz_dplyr, fz_output)
+
+
+  # States from output
+  shock_s1_output <- results_nl$xy_data_sets[[1]]$shock_s1
+  shock_s2_output <- results_nl$xy_data_sets[[1]]$shock_s2
+
+  # Make lagged tibble based on data set
+  data_set_lag <- data_set %>%
+                  arrange(cross_section, time_section) %>%
+                  group_by(cross_section)              %>%
+                  mutate_at(vars(y), funs(dplyr::lag(., 1))) %>%
+                  ungroup()
+
+  lag_indicator <-  !is.na(data_set_lag$y)
+
+  # Manual states
+  shock_s1_manual <- data_set$x_1*(1  - fz_dplyr)
+  shock_s1_manual <- shock_s1_manual[lag_indicator]
+
+  shock_s2_manual <- data_set$x_1*(fz_dplyr)
+  shock_s2_manual <- shock_s2_manual[lag_indicator]
+
+  # Test whether states of shock variable are correctly estimated
+  testthat::expect_equal(as.numeric(shock_s1_output), shock_s1_manual)
+  testthat::expect_equal(as.numeric(shock_s2_output), shock_s2_manual)
+
+
+
+})